[broker] Send total balances to trading page (bis)
This commit sends the total balances to the trading page
using websocket balance updates. Compared to !1861, this
commit does not break the front-end and also does not
send two messages back to back for each kind of balances.
We only send a single protobuf message by adding a field:
`total`, to the Balances message. This is more efficient
on the wire and makes future refactoring of broker easier.

Signed-off-by: AlexandreB <alexandreb@lich.io>