Move WebAuthenticationResult protobuf type into the generic client package
With the addition of the trigger system, which will have it's own
websocket component, there will be another service using the wsutil
package. The wsutil package currently contains a significant amount of
broker-specific code, which makes using the wsutil package for anything
except broker difficult. This is a (somewhat) small change to move the
WebAuthenticationResult type to client (the generic proto package), and
move the code to send it from wsutil into broker. This way trigger will
be able to use WebAuthenticationResult without needing to muck around
with broker types.

Doing this required manually modifying the proto.lock, which got angry
that the WebAuthenticationResult field type in BrokerUpdateMessage
changed. The "new" type is exactly the same as the old, just in a
different place, so protobufs should treat it exactly the same.